+++ Chirp.sln
@@ -11,6 +11,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{554816A0-C1B
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Razor", "src\Chirp.Razor\Chirp.Razor.csproj", "{EF1C5549-3FBB-47E2-A371-EDBCB1F9C85C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Core", "src\Chirp.Core\Chirp.Core.csproj", "{5DF1FF93-86EC-4B64-8777-5320C5772FAF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -28,9 +30,14 @@ Global
 		{EF1C5549-3FBB-47E2-A371-EDBCB1F9C85C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EF1C5549-3FBB-47E2-A371-EDBCB1F9C85C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EF1C5549-3FBB-47E2-A371-EDBCB1F9C85C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5DF1FF93-86EC-4B64-8777-5320C5772FAF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5DF1FF93-86EC-4B64-8777-5320C5772FAF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5DF1FF93-86EC-4B64-8777-5320C5772FAF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5DF1FF93-86EC-4B64-8777-5320C5772FAF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{6EAA5D16-9E9D-4928-988F-6D353D627515} = {0B9FF8D0-AF47-4795-99AC-42B904AFBD07}
 		{EF1C5549-3FBB-47E2-A371-EDBCB1F9C85C} = {554816A0-C1B7-4D0E-9DBB-56E5BEC854E1}
+		{5DF1FF93-86EC-4B64-8777-5320C5772FAF} = {554816A0-C1B7-4D0E-9DBB-56E5BEC854E1}
 	EndGlobalSection
 EndGlobal
