+++ src/Chirp.Core/DTOs/CreateCheepDTO.cs
@@ -1,0 +1,8 @@
+public class CreateCheepDTO
+{
+	public Guid CheepGuid { get; set; }
+	public required string Text { get; set; }
+	public required int AuthorId { get; set; }
+	public required string Name { get; set; }
+	public required string Email { get; set; }
+}
