+++ src/Chirp.Core/IAuthorRepository.cs
@@ -1,0 +1,7 @@
+namespace Chirp.Core;
+public interface IAuthorRepository
+{
+	public List<AuthorDTO> GetAuthorByName(string name);
+	public List<AuthorDTO> GetAuthorByEmail(string email);
+	public void CreateNewAuthor(Guid id, string name, string email);
+}
