+++ src/Chirp.Core/ICheepRepository.cs
@@ -1,0 +1,8 @@
+namespace Chirp.Core;
+public interface ICheepRepository
+{
+	public List<CheepDTO> GetCheeps(int page);
+	public List<CheepDTO> GetCheepsFromAuthor(int page, string author);
+	public void CreateNewCheep(CreateCheepDTO createCheepDTO);
+
+}
