+++ src/Chirp.Razor/Pages/UserTimeline.cshtml.cs
@@ -1,4 +1,4 @@
-﻿using Chirp.Razor.Repositories;
+﻿using Chirp.Core;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
