+++ src/Chirp.Razor/Repositories/AuthorRepository.cs
@@ -1,12 +1,6 @@
-﻿namespace Chirp.Razor.Repositories
+﻿using Chirp.Core;
+namespace Chirp.Razor.Repositories
 {
-	public interface IAuthorRepository
-	{
-		public List<AuthorDTO> GetAuthorByName(string name);
-		public List<AuthorDTO> GetAuthorByEmail(string email);
-		public void CreateNewAuthor(Guid id, string name, string email);
-	}
-
 	public class AuthorRepository : IAuthorRepository
 	{
 		private readonly ChirpDBContext _context;
